@@ -8,34 +8,34 @@
     ],
     "license": "MIT",
     "require": {
-        "php": "^7.2.5",
-        "bensampo/laravel-enum": "^2.1",
+        "php": "^7.3.0",
+        "bensampo/laravel-enum": "^3.0",
         "brian2694/laravel-toastr": "^5.54",
         "components/flag-icon-css": "^3.5",
         "dompdf/dompdf": "^0.8.5",
         "fideloper/proxy": "^4.2",
         "fruitcake/laravel-cors": "^1.0",
-        "graylog2/gelf-php": "1.5",
-        "guzzlehttp/guzzle": "^6.3",
+        "graylog2/gelf-php": "^1.5",
+        "guzzlehttp/guzzle": "^7.0.1",
         "itsjavi/bootstrap-colorpicker": "^3.2",
         "jeroennoten/laravel-adminlte": "^3.2",
         "laravel-admin-ext/daterangepicker": "^1.0",
-        "laravel/framework": "^7.0",
+        "laravel/framework": "^8.0",
         "laravel/tinker": "^2.0",
-        "laravel/ui": "^2.2",
+        "laravel/ui": "^3.0",
         "laravelcollective/html": "^6.1",
         "maatwebsite/excel": "^3.1",
-        "mockery/mockery": "1.3.1",
+        "mockery/mockery": "^1.3.1",
         "owen-it/laravel-auditing": "^10.0",
         "predis/predis": "^1.0",
         "yajra/laravel-datatables-oracle": "^9.10",
-         "symfony/http-kernel": "5.1.5"
+         "symfony/http-kernel": "^5.1.5"
     },
     "require-dev": {
-        "facade/ignition": "^2.0",
+        "facade/ignition": "^2.3.6",
         "fzaninotto/faker": "^1.9.1",
-        "nunomaduro/collision": "^4.1",
-        "phpunit/phpunit": "^8.5"
+        "nunomaduro/collision": "^5.0",
+        "phpunit/phpunit": "^9.0"
     },
     "config": {
         "optimize-autoloader": true,
@@ -49,12 +49,10 @@
     },
     "autoload": {
         "psr-4": {
-            "App\\": "app/"
-        },
-        "classmap": [
-            "database/seeds",
-            "database/factories"
-        ]
+            "App\\": "app/",
+            "Database\\Factories\\": "database/factories/",
+            "Database\\Seeders\\": "database/seeders/"
+        }
     },
     "autoload-dev": {
         "psr-4": {
