@@ -15,34 +15,33 @@
         }
     ],
     "require": {
-        "php": "^7.2.5",
+        "php": "^7.3",
         "ext-json": "*",
         "ext-mbstring": "*",
         "ext-openssl": "*",
         "doctrine/inflector": "^1.4|^2.0",
-        "dragonmantank/cron-expression": "^2.0",
+        "dragonmantank/cron-expression": "^3.0",
         "egulias/email-validator": "^2.1.10",
         "league/commonmark": "^1.3",
         "league/flysystem": "^1.0.34",
         "monolog/monolog": "^2.0",
         "nesbot/carbon": "^2.17",
-        "opis/closure": "^3.1",
+        "opis/closure": "^3.5.3",
         "psr/container": "^1.0",
         "psr/simple-cache": "^1.0",
-        "ramsey/uuid": "^3.7|^4.0",
+        "ramsey/uuid": "^4.0",
         "swiftmailer/swiftmailer": "^6.0",
-        "symfony/console": "^5.0",
-        "symfony/error-handler": "^5.0",
-        "symfony/finder": "^5.0",
-        "symfony/http-foundation": "^5.0",
-        "symfony/http-kernel": "^5.0",
-        "symfony/mime": "^5.0",
-        "symfony/polyfill-php73": "^1.17",
-        "symfony/process": "^5.0",
-        "symfony/routing": "^5.0",
-        "symfony/var-dumper": "^5.0",
+        "symfony/console": "^5.1",
+        "symfony/error-handler": "^5.1",
+        "symfony/finder": "^5.1",
+        "symfony/http-foundation": "^5.1",
+        "symfony/http-kernel": "^5.1",
+        "symfony/mime": "^5.1",
+        "symfony/process": "^5.1",
+        "symfony/routing": "^5.1",
+        "symfony/var-dumper": "^5.1",
         "tijsverkoyen/css-to-inline-styles": "^2.2.2",
-        "vlucas/phpdotenv": "^4.0",
+        "vlucas/phpdotenv": "^5.2",
         "voku/portable-ascii": "^1.4.8"
     },
     "replace": {
