@@ -14,11 +14,13 @@
         }
     ],
     "require": {
-        "php": "^7.2.5",
-        "illuminate/contracts": "^7.0",
-        "illuminate/http": "^7.0",
-        "illuminate/queue": "^7.0",
-        "illuminate/support": "^7.0"
+        "php": "^7.3",
+        "illuminate/collections": "^8.0",
+        "illuminate/contracts": "^8.0",
+        "illuminate/http": "^8.0",
+        "illuminate/macroable": "^8.0",
+        "illuminate/queue": "^8.0",
+        "illuminate/support": "^8.0"
     },
     "autoload": {
         "psr-4": {
@@ -27,13 +29,13 @@
     },
     "extra": {
         "branch-alias": {
-            "dev-master": "7.x-dev"
+            "dev-master": "8.x-dev"
         }
     },
     "suggest": {
-        "illuminate/console": "Required to use the auth:clear-resets command (^7.0).",
-        "illuminate/queue": "Required to fire login / logout events (^7.0).",
-        "illuminate/session": "Required to use the session based guard (^7.0)."
+        "illuminate/console": "Required to use the auth:clear-resets command (^8.0).",
+        "illuminate/queue": "Required to fire login / logout events (^8.0).",
+        "illuminate/session": "Required to use the session based guard (^8.0)."
     },
     "config": {
         "sort-packages": true
