@@ -14,9 +14,11 @@
         }
     ],
     "require": {
-        "php": "^7.2.5",
-        "illuminate/contracts": "^7.0",
-        "illuminate/support": "^7.0"
+        "php": "^7.3",
+        "illuminate/collections": "^8.0",
+        "illuminate/contracts": "^8.0",
+        "illuminate/macroable": "^8.0",
+        "illuminate/support": "^8.0"
     },
     "autoload": {
         "psr-4": {
@@ -25,15 +27,15 @@
     },
     "extra": {
         "branch-alias": {
-            "dev-master": "7.x-dev"
+            "dev-master": "8.x-dev"
         }
     },
     "suggest": {
         "ext-memcached": "Required to use the memcache cache driver.",
-        "illuminate/database": "Required to use the database cache driver (^7.0).",
-        "illuminate/filesystem": "Required to use the file cache driver (^7.0).",
-        "illuminate/redis": "Required to use the redis cache driver (^7.0).",
-        "symfony/cache": "Required to PSR-6 cache bridge (^5.0)."
+        "illuminate/database": "Required to use the database cache driver (^8.0).",
+        "illuminate/filesystem": "Required to use the file cache driver (^8.0).",
+        "illuminate/redis": "Required to use the redis cache driver (^8.0).",
+        "symfony/cache": "Required to PSR-6 cache bridge (^5.1)."
     },
     "config": {
         "sort-packages": true
