@@ -1,15 +1,28 @@
 <?php
+namespace Database\Factories;
 
-/** @var \Illuminate\Database\Eloquent\Factory $factory */
+use Illuminate\Database\Eloquent\Factories\Factory;
+use Illuminate\Support\Str;
 
 use App\Region;
-use Faker\Generator as Faker;
 
-$factory->define(Region::class, function (Faker $faker) {
+
+class RegionFactory extends Factory
+{
+  /**
+   * The name of the factory's corresponding model.
+   *
+   * @var string
+   */
+  protected $model = Region::class;
+
+  public function definition()
+  {
     return [
         'name' => 'Hessischer Basketball Verband',
         'code' => 'HBV',
         'hq' => '',
         'created_at' => now()
     ];
-});
+  }
+}
